+++ src/rer/solrpush/restapi/services/solr_search/solr_search_handler.py
@@ -6,7 +6,6 @@ from Products.CMFCore.utils import getToolByName
 from rer.solrpush.restapi.services.solr_search.batch import SolrHypermediaBatch
 from rer.solrpush.utils import search as solr_search
 
-
 DEFAULT_METADATA_FIELDS = set(
     ["@id", "@type", "description", "review_state", "title"]
 )